Added Location Module Implementation
Added Location Module Implementation - Location module (php file with javascript/php depending on the request). When no address is provided a javascript code is activated and returns a JSON with the client's adress. When a request is made and an address provided, a php code is activated and returns a JSON with that address details.